Last Minute Revision Toolkit by Vishwas (#453)
* Create OddEvenLL.py

* Create DeleteMiddleNode.py

* Create count_hills&valleys.py

Given a 0-indexed integer array nums. An index i is part of a hill in nums if the closest non-equal neighbors of i are smaller than nums[i]. Similarly, an index i is part of a valley in nums if the closest non-equal neighbors of i are larger than nums[i]. Adjacent indices i and j are part of the same hill or valley if nums[i] == nums[j].

Note that for an index to be part of a hill or valley, it must have a non-equal neighbor on both the left and right of the index.

Return the number of hills and valleys in nums.

* Update count_hills&valleys.py by Vishwas

* Update DeleteMiddleNode.py by Vishwas

* Update OddEvenLL.py by Vishwas

* Update count_hills&valleys.py by Vishwas